@@ -1,6 +1,6 @@
 module github.com/amimof/blipblop
 
-go 1.17
+go 1.19
 
 require (
 	github.com/containerd/go-cni v1.1.4
